# php/crud-example/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Tareas",
        "version": "1.0"
    },
    "paths": {
        "/tasks": {
            "get": {
                "summary": "Obtener lista de tareas",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "Lista de tareas"
                    }
                }
            },
            "post": {
                "summary": "Crear nueva tarea",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "summary": "Obtener tarea por ID",
                "operationId": "fd3677b5b24695d7b801715869b1357c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea obtenida"
                    }
                }
            },
            "put": {
                "summary": "Actualizar tarea",
                "operationId": "85377ff0e935f093dd75ea6886923dbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar tarea",
                "operationId": "5ed5dfd4c645e9398f4282969bfe6478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada"
                    }
                }
            }
        }
    }
}